Helps address #17 problems.
These changes allow Often to run when bcrypt module is not installed. This change is intended only for development model and not the production server. As such, when development mode is off bcrypt will always be used even if not installed.

A new module was added to help general crypt methods. Bcrypt is loaded if installed, otherwise node's default crypt is loaded and SHA1 hashing is used instead. `server.js` was modified to set a flag to allow this behavior.